Added person tracking and twilio test code
The api key for twilio is stored outside of this git repo for security. Texting and then having that message show up on the display works! My plan now is to have users text the twilio number, have opencv startup the camera tracking and then start pong/brick breaker or whatever.